Implemented private messages
There's currently two bugs. 1) You can find yourself and send messages to yourself (we can use it as a feature, but to make it a feature we should fix a second bug). 2) The user receives a message even though he's is not in the selected chatroom. How can we fix that? I have no clue.

TODO -- Fix them.